Split Vue.Router routes
