+++ website/client/src/router/shared-route-imports.js
@@ -1,0 +1,3 @@
+export const NotFoundPage = () => import(/* webpackChunkName: "not-found" */'@/components/404');
+
+export const ProfilePage = () => import(/* webpackChunkName: "user" */'@/components/userMenu/profilePage');
